Revert "Refactor token login (#7380)"
This reverts commit 36c1b57a
